+++ app/src/main/java/com/example/playlistmaker/media/domain/models/Playlist.kt
@@ -10,5 +10,5 @@ data class Playlist(
     val description: String?,
     val coverUri: String?,
     val trackIds: List<Long> = emptyList(),
-    val playlistSize: Long = 0L
+    val playlistSize: Int = 0
 ) : Parcelable
